+++ src/plugins/blocksmith/components/blocksmith/selector.js
@@ -1,10 +1,16 @@
+import getBlocksmith from '@triniti/cms/plugins/blocksmith/selectors/getBlocksmith';
 import getCopiedBlock from '../../selectors/getCopiedBlock';
 
 /**
  * @param {Object} state - The entire redux state.
+ * @param {Object} ownProps
  *
  * @returns {Object}
  */
-export default (state) => ({
-  copiedBlock: getCopiedBlock(state),
-});
+export default (state, { formName }) => {
+  const blocksmithState = getBlocksmith(state, formName);
+  return {
+    copiedBlock: getCopiedBlock(state),
+    editorState: blocksmithState && blocksmithState.editorState ? blocksmithState.editorState : null,
+  };
+};