+++ src/plugins/blocksmith/utils/getBlockForKey.js
@@ -7,13 +7,6 @@ import normalizeKey from './normalizeKey';
  * @param {ContentState} contentState - a state instance of a DraftJs Editor
  * @param {string}       key          - a block, a block index, or a block key
  *
- * @returns {EditorState} an EditorState instance
+ * @returns {?ContentBlock}
  */
-
-export default (contentState, key) => {
-  if (typeof key !== 'string') {
-    throw new Error(`key ['${key}'] is not a string`);
-  } else {
-    return contentState.getBlockForKey(normalizeKey(key));
-  }
-};
+export default (contentState, key) => contentState.getBlockForKey(normalizeKey(key));