+++ src/plugins/blocksmith/utils/normalizeKey.test.js
@@ -15,11 +15,5 @@ test('Blocksmith:util:normalizeKey', (t) => {
   actual = normalizeKey('asdf');
   t.equal(actual, expected, 'It should return an api key when given an api key');
 
-  t.throws(
-    () => normalizeKey('@#D^!@'),
-    new Error('Key [@#D^!@] does not appear to be a valid ContentBlock key.'),
-    'It should throw an error when provided with an invalid key.',
-  );
-
   t.end();
 });