+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
 
+## vN.N.N
+* Blocksmith error reporting & recovery
+
+
 ## v0.9.4
 * Remove Link to Google.com from Poster Image In Gallery Block Creator
 * Display error notification when CMS web socket loses connection.